+++ NEWS.md
@@ -1,3 +1,6 @@
+- new function `ml_fit_dss` with an implementation very close to the paper by
+  Deville et al. (1993); implementation in the `laeken` package
+
 v0.0-11.2 (2014-11-15)
 ===